# contracts/openapi/v0.2/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vehicle System Release Quality Gate API",
    "version": "0.2.0-draft.2",
    "description": "Machine-executable V0.2 draft. V0.1 Core Contract semantics remain authoritative."
  },
  "servers": [{ "url": "https://vsrqg.example" }],
...
  "paths": {
...
    "/api/v1/test-runs": {
      "post": {
        "operationId": "createTestRun",
        "tags": ["Test"],
        "x-permission": "test:execute",
        "x-idempotency-required": true,
        "security": [{ "oidc": ["test:execute"] }],
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": { "$ref": "#/components/requestBodies/CreateTestRun" },
        "responses": { "201": { "$ref": "#/components/responses/ResourceCreated" }, "409": { "$ref": "#/components/responses/Problem" }, "default": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/api/v1/test-runs/{id}:cancel": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "post": {
        "operationId": "cancelTestRun",
        "tags": ["Test"],
        "x-permission": "test:execute",
        "x-idempotency-required": true,
        "security": [{ "oidc": ["test:execute"] }],
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": { "$ref": "#/components/requestBodies/Reason" },
        "responses": { "202": { "$ref": "#/components/responses/Accepted" }, "409": { "$ref": "#/components/responses/Problem" }, "default": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/api/v1/test-runs/{id}/results": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "get": {
        "operationId": "getTestRunResults",
        "tags": ["Test"],
        "x-permission": "test:read",
        "security": [{ "oidc": ["test:read"] }],
        "responses": { "200": { "$ref": "#/components/responses/ResourceList" }, "default": { "$ref": "#/components/responses/Problem" } }
      }
    },
...
    "/agent-api/v1/agents/{id}:heartbeat": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "post": {
        "operationId": "heartbeatAgent",
        "tags": ["Agent"],
        "x-permission": "agent:heartbeat",
        "x-idempotency-required": true,
        "security": [{ "agentMtls": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": { "$ref": "#/components/requestBodies/AgentHeartbeat" },
        "responses": { "200": { "$ref": "#/components/responses/Resource" }, "409": { "$ref": "#/components/responses/Problem" }, "default": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/agent-api/v1/agents/{id}/commands:poll": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "post": {
        "operationId": "pollAgentCommands",
        "tags": ["Agent"],
        "x-permission": "agent:poll",
        "x-idempotency-required": true,
        "security": [{ "agentMtls": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": { "$ref": "#/components/requestBodies/AgentPoll" },
        "responses": { "200": { "$ref": "#/components/responses/AgentCommand" }, "default": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/agent-api/v1/commands/{commandId}:ack": {
      "parameters": [{ "$ref": "#/components/parameters/CommandId" }],
      "post": {
        "operationId": "acknowledgeAgentCommand",
        "tags": ["Agent"],
        "x-permission": "agent:execute",
        "x-idempotency-required": true,
        "security": [{ "agentMtls": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": { "$ref": "#/components/requestBodies/AgentAck" },
        "responses": { "200": { "$ref": "#/components/responses/Resource" }, "409": { "$ref": "#/components/responses/Problem" }, "default": { "$ref": "#/components/responses/Problem" } }
      }
    },
...
    "/agent-api/v1/attempts/{attemptId}/context": {
      "parameters": [{ "$ref": "#/components/parameters/AttemptId" }],
      "get": {
        "operationId": "getAgentExecutionContext",
        "tags": ["Agent"],
        "description": "Implemented by Task 3. Only the assigned Agent may read a dispatched Attempt context. Attempt IDs are standard lowercase UUIDs. No credentials, host paths or raw device serial numbers.",
        "x-permission": "agent:execute",
        "x-idempotency-required": false,
        "security": [{ "agentMtls": [] }],
        "responses": {
          "200": { "description": "Immutable assigned execution context", "content": { "application/json": { "schema": { "$ref": "../../../schemas/v0.2/agent-execution-context.schema.json" } } } },
          "default": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "oidc": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://identity.example/authorize",
            "tokenUrl": "https://identity.example/token",
            "scopes": {
              "release:create": "Create Release identities",
              "release:read": "Read Releases",
              "release:approve": "Approve Releases",
              "manifest:write": "Register and validate Manifests",
              "manifest:lock": "Lock authoritative Manifests",
              "issue:sync": "Synchronize Issue sources",
              "issue:configure": "Configure Issue source mapping profiles",
              "issue:read": "Read Issue synchronization runs",
              "issue:snapshot": "Create Issue snapshots",
              "traceability:read": "Read traceability",
              "traceability:verify": "Verify traceability snapshots",
              "test:execute": "Create and cancel Test Runs",
              "test:read": "Read Test Results",
              "evidence:read": "Read Evidence metadata and non-HIGH payloads",
              "evidence:read:sensitive": "Read HIGH Evidence payloads",
              "rule:write": "Author Rule Sets",
              "rule:publish": "Publish Rule Sets",
              "quality:evaluate": "Request deterministic evaluations",
              "quality:read": "Read Quality Results",
              "quality:override": "Create governance decisions"
            }
          }
        }
      },
      "serviceOauth": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "https://identity.example/token",
            "scopes": {
              "traceability:ingest": "Ingest Traceability facts as a service identity"
            }
          }
...
    "parameters": {
      "IdempotencyKey": { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "minLength": 1, "maxLength": 128 } },
...
      "CommandId": { "name": "commandId", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/OpaqueId" } },
      "AttemptId": { "name": "attemptId", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/OpaqueId" } },
      "Id": { "name": "id", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/OpaqueId" } }
    },
    "schemas": {
      "OpaqueId": { "type": "string", "minLength": 1, "maxLength": 128 },
      "GenericResource": { "type": "object", "additionalProperties": true },
      "GenericList": {
        "type": "object",
        "required": ["items"],
        "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/GenericResource" } }, "nextCursor": { "type": ["string", "null"] } },
        "additionalProperties": false
      },
      "Problem": {
        "type": "object",
        "required": ["type", "title", "status", "code", "requestId"],
        "properties": {
          "type": { "type": "string", "format": "uri" },
          "title": { "type": "string" },
          "status": { "type": "integer", "minimum": 400, "maximum": 599 },
          "code": { "type": "string", "pattern": "^[A-Z][A-Z0-9_]{2,63}$" },
          "detail": { "type": "string" },
          "instance": { "type": "string" },
          "requestId": { "$ref": "#/components/schemas/OpaqueId" },
          "violations": { "type": "array", "items": { "$ref": "#/components/schemas/GenericResource" } }
        },
        "additionalProperties": false
      },
...
      "Reason": {
        "type": "object",
        "required": ["reason"],
        "properties": { "reason": { "type": "string", "minLength": 1, "maxLength": 1000 } },
        "additionalProperties": false
      },
...
      "CreateTestRunRequest": {
        "type": "object",
        "required": ["releaseId", "testPlan", "deviceSelector"],
        "properties": {
          "releaseId": { "$ref": "#/components/schemas/OpaqueId" },
          "testPlan": {
            "type": "object",
            "required": ["planId", "version"],
            "properties": { "planId": { "$ref": "#/components/schemas/OpaqueId" }, "version": { "type": "integer", "minimum": 1 } },
            "additionalProperties": false
          },
          "deviceSelector": {
            "type": "object",
            "required": ["vehicle", "requiredCapabilities"],
            "properties": { "vehicle": { "type": "string", "minLength": 1 }, "requiredCapabilities": { "type": "array", "uniqueItems": true, "items": { "type": "string", "minLength": 1 } } },
            "additionalProperties": false
          }
        },
        "additionalProperties": false
      },
...
    },
    "requestBodies": {
...
      "Reason": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Reason" } } } },
...
      "CreateTestRun": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateTestRunRequest" } } } },
...
      "AgentHeartbeat": { "required": true, "content": { "application/json": { "schema": { "$ref": "../../../schemas/v0.2/agent-protocol.schema.json#/$defs/heartbeatRequest" } } } },
      "AgentPoll": { "required": true, "content": { "application/json": { "schema": { "$ref": "../../../schemas/v0.2/agent-protocol.schema.json#/$defs/pollRequest" } } } },
      "AgentAck": { "required": true, "content": { "application/json": { "schema": { "$ref": "../../../schemas/v0.2/agent-protocol.schema.json#/$defs/ackRequest" } } } },
...
    },
    "responses": {
...
      "Resource": { "description": "Resource response", "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GenericResource" } } } },
...
      "ResourceCreated": { "description": "Resource created", "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GenericResource" } } } },
...
      "Accepted": { "description": "Asynchronous work accepted", "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GenericResource" } } } },
      "ResourceList": { "description": "Cursor-page response", "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GenericList" } } } },
      "Problem": { "description": "RFC 9457 Problem Details", "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } }, "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } },
...
      "AgentCommand": { "description": "Zero or one command for the negotiated protocol", "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } }, "content": { "application/json": { "schema": { "oneOf": [{ "$ref": "../../../schemas/v0.2/agent-protocol.schema.json#/$defs/commandEnvelope" }, { "type": "null" }] } } } }
    },
    "headers": {
      "RequestId": { "required": true, "schema": { "$ref": "#/components/schemas/OpaqueId" } }
    }
  }
}